build static C++ libs then link them to py bindings
