+++ CMakeLists.txt
@@ -27,22 +27,64 @@ if("${RF24_DRIVER}" STREQUAL "RPi" OR "${RF24_DRIVER}" STREQUAL "SPIDEV")
     list(APPEND RF24_DRIVER_SOURCES src/linux/gpio.h)
 endif()
 
+### Build C++ RF24 stack as static libraries (for linking to pyRF24*)
+
+################################# RF24 C++ #########################
+add_library(cpp_rf24 STATIC
+    RF24/RF24.cpp
+    ${RF24_DRIVER_SOURCES}
+)
+if(SUPPLEMENT_LINUX_GPIO_H)
+    target_include_directories(cpp_rf24 PUBLIC src)
+endif()
+target_include_directories(cpp_rf24 PUBLIC utility)
+if(NOT "${RF24_LINKED_DRIVER}" STREQUAL "")
+    if("${RF24_DRIVER}" STREQUAL "wiringPi")
+        target_link_libraries(cpp_rf24 PRIVATE rt crypt ${RF24_LINKED_DRIVER})
+    else()
+        target_link_libraries(cpp_rf24 PRIVATE ${RF24_LINKED_DRIVER})
+    endif()
+endif()
+apply_flags(cpp_rf24)
+
+################################# RF24Network C++ #########################
+add_library(cpp_rf24_network STATIC
+    RF24Network/RF24Network.cpp
+)
+# don't let source look for an installed RF24 lib
+target_compile_definitions(cpp_rf24_network PUBLIC USE_RF24_LIB_SRC)
+target_include_directories(cpp_rf24_network PUBLIC
+    RF24
+    RF24/utility
+    RF24/utility/${RF24_DRIVER}
+    RF24Network
+)
+target_link_libraries(cpp_rf24_network PRIVATE cpp_rf24)
+apply_flags(cpp_rf24_network)
+
+################################# RF24Mesh C++ #########################
+add_library(cpp_rf24_mesh STATIC
+    RF24Mesh/RF24Mesh.cpp
+)
+# don't let source look for an installed RF24 lib
+target_compile_definitions(cpp_rf24_mesh PUBLIC USE_RF24_LIB_SRC)
+target_include_directories(cpp_rf24_mesh PUBLIC
+    RF24
+    RF24/utility
+    RF24/utility/${RF24_DRIVER}
+    RF24Network
+    RF24Mesh
+)
+target_link_libraries(cpp_rf24_mesh PRIVATE cpp_rf24_network)
+apply_flags(cpp_rf24_mesh)
+
 ################################# RF24 #############################
 pybind11_add_module(rf24 src/pyRF24.cpp)
+target_link_libraries(rf24 PRIVATE cpp_rf24)
 target_include_directories(rf24 PUBLIC
     RF24
     RF24/utility
     RF24/utility/${RF24_DRIVER}
-)
-if(SUPPLEMENT_LINUX_GPIO_H)
-    target_include_directories(rf24 PUBLIC src)
-endif()
-target_sources(rf24 PUBLIC
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/nRF24L01.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24_config.h
-    ${RF24_DRIVER_SOURCES}
 )
 apply_flags(rf24)
 
@@ -51,25 +93,13 @@ apply_flags(rf24)
 pybind11_add_module(rf24_network src/pyRF24Network.cpp)
 target_include_directories(rf24_network PUBLIC
     RF24
-    RF24Network
     RF24/utility
     RF24/utility/${RF24_DRIVER}
+    RF24Network
 )
-if(SUPPLEMENT_LINUX_GPIO_H)
-    target_include_directories(rf24_network PUBLIC src)
-endif()
 # don't let source look for an installed RF24 lib
 target_compile_definitions(rf24_network PUBLIC USE_RF24_LIB_SRC)
-target_sources(rf24_network PUBLIC
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/nRF24L01.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24_config.h
-    ${RF24_DRIVER_SOURCES}
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Network/RF24Network.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Network/RF24Network.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Network/RF24Network_config.h
-)
+target_link_libraries(rf24_network PRIVATE cpp_rf24_network)
 apply_flags(rf24_network)
 
 
@@ -78,30 +108,20 @@ apply_flags(rf24_network)
 pybind11_add_module(rf24_mesh src/pyRF24Mesh.cpp)
 target_include_directories(rf24_mesh PUBLIC
     RF24
-    RF24Network
-    RF24Mesh
     RF24/utility
     RF24/utility/${RF24_DRIVER}
+    RF24Network
+    RF24Mesh
 )
-if(SUPPLEMENT_LINUX_GPIO_H)
-    target_include_directories(rf24_mesh PUBLIC src)
-endif()
+target_link_libraries(rf24_mesh PRIVATE cpp_rf24_mesh)
 # don't let source look for an installed RF24 lib
 target_compile_definitions(rf24_mesh PUBLIC USE_RF24_LIB_SRC)
-target_sources(rf24_mesh PUBLIC
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/nRF24L01.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24/RF24_config.h
-    ${RF24_DRIVER_SOURCES}
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Network/RF24Network.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Network/RF24Network.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Network/RF24Network_config.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Mesh/RF24Mesh.h
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Mesh/RF24Mesh.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/RF24Mesh/RF24Mesh_config.h
-)
-apply_flags(rf24_mesh)
+# target_sources(rf24_mesh PUBLIC
+#     ${CMAKE_CURRENT_LIST_DIR}/RF24Mesh/RF24Mesh.h
+#     ${CMAKE_CURRENT_LIST_DIR}/RF24Mesh/RF24Mesh.cpp
+#     ${CMAKE_CURRENT_LIST_DIR}/RF24Mesh/RF24Mesh_config.h
+# )
+# apply_flags(rf24_mesh)
 
 
 ################################ INSTALL RULES ####################################
@@ -112,4 +132,4 @@ install(TARGETS rf24_network DESTINATION .)
 install(TARGETS rf24_mesh DESTINATION .)
 
 # uncomment to show compiler args used in build logs
-# set(CMAKE_VERBOSE_MAKEFILE ON)
+set(CMAKE_VERBOSE_MAKEFILE ON)
